backend/requirements.txt
fastapi
uvicorn
sqlalchemy
pydantic
python-multipart
python-jose[cryptography]
passlib[bcrypt]
argon2-cffi
httpx
pytest
psycopg2-binary
python-dotenv
google-generativeai
